restricted update,delete from another users

+++ base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.contrib.auth.models import User
@@ -13,8 +14,7 @@ def login_page(request):
 
         # print(username,password)
         try:
-            user = User.objects.get(username=username)
-            print(user)
+            User.objects.get(username=username)
 
         except:
             print("NOT exist")
@@ -67,22 +67,30 @@ def create_room(request):
     context = {"form": form}
     return render(request, "room_form.html", context)
 
+
 @login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("you cannot update another user room")
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
             return redirect("home")
 
+
     context = {"form": form}
     return render(request, "room_form.html", context)
+
 
 @login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("you cannot delete another user room")
     if request.method == "POST":
         room.delete()
         return redirect("home")
